+++ src/components/form/formComponent.jsx
@@ -1,10 +1,8 @@
 import React from 'react';
 
-import isFunction from 'lodash-es/isFunction';
-import some from 'lodash-es/some';
-import compact from 'lodash-es/compact';
-import debounce from 'lodash-es/debounce';
 import styled from 'styled-components';
+
+import { debounce, find } from '../../helpers/utils';
 
 import { PrimaryButton } from './buttonComponent';
 import { Error } from '../miscComponent';
@@ -33,13 +31,7 @@ export function createForm(config) {
         constructor(props) {
             super(props);
 
-            this.allFields = compact(isFunction(props.fields) ? props.fields(props) : props.fields).map(f => (
-                !f.staticContent ? f.create({ i18n: props.i18n, showLabel: props.showLabels }) : f
-            ));
-
-            this.inputFields = this.allFields.filter(f => !f.staticContent);
-
-            this.fieldByKey = this.inputFields.reduce((acc, field) => ({ ...acc, [field.key]: field }), {});
+            this.updateFields(props)
 
             this.state = {
                 isLoading: false,
@@ -47,6 +39,21 @@ export function createForm(config) {
                 errorMessage: null,
                 fields: this.applyModel(props.initialModel)
             };
+        }
+
+        updateFields(props) {
+            this.allFields = (typeof props.fields === 'function' ? props.fields(props) : props.fields ?? []).filter(x => !!x).map(f => (
+                !f.staticContent ? f.create({ i18n: props.i18n, showLabel: props.showLabels }) : f
+            ));
+
+            this.inputFields = this.allFields.filter(f => !f.staticContent);
+
+            this.fieldByKey = this.inputFields.reduce((acc, field) => ({ ...acc, [field.key]: field }), {});
+        }
+
+
+        UNSAFE_componentWillReceiveProps(props) {
+            this.updateFields(props)
         }
 
         componentWillUnmount() {
@@ -99,7 +106,7 @@ export function createForm(config) {
                 const currentState = prevState.fields[fieldName];
                 const newState = {
                     ...currentState,
-                    ...(isFunction(stateUpdate) ? stateUpdate(currentState) : stateUpdate)
+                    ...(typeof stateUpdate === 'function' ? stateUpdate(currentState) : stateUpdate)
                 };
                 const newFields = {
                     ...prevState.fields,
@@ -127,7 +134,7 @@ export function createForm(config) {
                 };
 
                 return {
-                    hasErrors: !!validation.error || some(newFields, ({ validation = {} }) => validation.error),
+                    hasErrors: !!validation.error || find(newFields, ({ validation = {} }) => validation.error) !== undefined,
                     fields: newFields
                 };
             });
@@ -203,7 +210,7 @@ export function createForm(config) {
         }
 
         render() {
-            const { submitLabel, allowWebAuthnLogin, i18n, fieldValidationDebounce } = this.props;
+            const { submitLabel, i18n, fieldValidationDebounce } = this.props;
             const { errorMessage, isLoading, fields } = this.state;
 
             return <Form noValidate onSubmit={this.handleSubmit}>
@@ -221,11 +228,10 @@ export function createForm(config) {
                     }) : field.staticContent)
                 }
                 {
-                    !allowWebAuthnLogin && <PrimaryButton disabled={isLoading}>
+                    <PrimaryButton disabled={isLoading}>
                         {i18n(submitLabel)}
                     </PrimaryButton>
                 }
-                {allowWebAuthnLogin && this.props.webAuthnButtons(isLoading, this.handleClick)}
             </Form>;
         }
     }