+++ types/identity-ui.d.ts
@@ -1,15 +1,15 @@
 /**
- * @reachfive/identity-ui - v1.25.2
- * Compiled Thu, 07 Mar 2024 14:43:01 UTC
+ * @reachfive/identity-ui - v1.26.0
+ * Compiled Tue, 09 Apr 2024 14:07:05 UTC
  *
  * Copyright (c) ReachFive.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  **/
+import { CSSProperties } from 'styled-components';
 import { Config as Config$1, RemoteSettings, ConsentVersions, CustomField, Client as Client$1, SessionInfo, AuthOptions, MFA, PasswordlessResponse, SingleFactorPasswordlessParams, Profile, UserConsent, DeviceCredential } from '@reachfive/identity-core';
 export { Config } from '@reachfive/identity-core';
-import CSS from 'csstype';
 import React$1 from 'react';
 import { PasswordlessParams } from '@reachfive/identity-core/es/main/oAuthClient';
 
@@ -30,8 +30,6 @@ type ConsentsVersions = { consentsVersions: Record<string, ConsentVersions> }
 type CustomFields = { customFields?: CustomField[] }
 
 type Config = Config$1 & RemoteSettings & ConsentsVersions & CustomFields
-
-declare const inputBtnFocusBoxShadow: (borderColor?: string) => string | undefined;
 
 declare module 'styled-components' {
     export interface DefaultTheme extends Theme {}
@@ -59,15 +57,15 @@ interface BaseTheme {
     /**
      * @default "#212529"
      */
-    headingColor: CSS.Color
+    headingColor: NonNullable<CSSProperties['color']>
     /**
      * @default "#495057"
      */
-    textColor: CSS.Color
+    textColor: NonNullable<CSSProperties['color']>
     /**
      * @default "#adb5bd"
      */
-    mutedTextColor: CSS.Color
+    mutedTextColor: NonNullable<CSSProperties['color']>
      /**
      * @default "3"
      */
@@ -75,7 +73,7 @@ interface BaseTheme {
     /**
      * @default "#ced4da "
      */
-    borderColor: CSS.Color
+    borderColor: NonNullable<CSSProperties['color']>
     /**
      * @default 1
      */
@@ -83,28 +81,28 @@ interface BaseTheme {
     /**
      * @default "#ffffff"
      */
-    backgroundColor: CSS.Color
+    backgroundColor: NonNullable<CSSProperties['color']>
     /**
      * The button and link default color.
      * @default "#229955"
      */
-    primaryColor: CSS.Color
+    primaryColor: NonNullable<CSSProperties['color']>
     /**
      * @default "#dc4e41"
      */
-    dangerColor: CSS.Color
+    dangerColor: NonNullable<CSSProperties['color']>
     /**
      * @default "#ffc107"
      */
-    warningColor: CSS.Color
+    warningColor: NonNullable<CSSProperties['color']>
     /**
      * @default "#229955"
      */
-    successColor: CSS.Color
+    successColor: NonNullable<CSSProperties['color']>
     /**
      * @default "#e9ecef"
      */
-    lightBackgroundColor: CSS.Color
+    lightBackgroundColor: NonNullable<CSSProperties['color']>
     /** Specifies the padding for the x axis. (left and right) */
     paddingX: number
     /** Specifies the padding for the y axis. (top and bottom) */
@@ -120,27 +118,27 @@ interface BaseTheme {
 }
 
 interface LinkTheme {
-    color: CSS.Color
-    decoration: CSS.TextDecorationLineProperty
-    hoverColor: CSS.Color
-    hoverDecoration: CSS.TextDecorationLineProperty
+    color: NonNullable<CSSProperties['color']>
+    decoration: NonNullable<CSSProperties['textDecoration']>
+    hoverColor: NonNullable<CSSProperties['color']>
+    hoverDecoration: NonNullable<CSSProperties['textDecoration']>
 }
 
 interface InputTheme {
-    color: CSS.Color
-    placeholderColor: CSS.Color
+    color: NonNullable<CSSProperties['color']>
+    placeholderColor: NonNullable<CSSProperties['color']>
     fontSize: number
     lineHeight: number
     paddingX: number
     paddingY: number
     borderRadius: number
-    borderColor: CSS.Color
+    borderColor: NonNullable<CSSProperties['color']>
     borderWidth: number
-    background: CSS.Color
-    disabledBackground: CSS.Color
-    boxShadow: string
-    focusBorderColor: CSS.Color
-    focusBoxShadow: typeof inputBtnFocusBoxShadow
+    background: NonNullable<CSSProperties['color']>
+    disabledBackground: NonNullable<CSSProperties['color']>
+    boxShadow: NonNullable<CSSProperties['boxShadow']>
+    focusBorderColor: NonNullable<CSSProperties['color']>
+    focusBoxShadow: (color?: CSSProperties['color']) => NonNullable<CSSProperties['boxShadow']>
     height: number
 }
 
@@ -148,7 +146,7 @@ interface ButtonTheme {
     /** Specifies the font-weight (such as normal, bold, or 900).
      * @default 'bold'
      */
-    fontWeight: CSS.FontWeightProperty
+    fontWeight: NonNullable<CSSProperties['fontWeight']>
     /** Specifies the font-size. */
     fontSize: number
      /** Specifies the line-height. */
@@ -157,12 +155,14 @@ interface ButtonTheme {
     paddingX: number
     /** Specifies the padding for the y axis. (top and bottom) */
     paddingY: number
+    /** Specifies the border-color. */
+    borderColor: NonNullable<CSSProperties['color']>,
     /** Specifies the border-radius. */
     borderRadius: number
     /** Specifies the border-width. */
     borderWidth: number
     /** Function that specifies the box shadow based on the border color. */
-    focusBoxShadow: typeof inputBtnFocusBoxShadow
+    focusBoxShadow: (color?: CSSProperties['color']) => NonNullable<CSSProperties['boxShadow']>
     /** Specifies the height. */
     height: number
 }
@@ -173,7 +173,7 @@ interface SocialButtonTheme {
      /** Boolean that specifies if the text is visible. */
     textVisible: boolean
     /** Specifies the font-weight (such as normal, bold, or 900). */
-    fontWeight: CSS.FontWeightProperty
+    fontWeight: NonNullable<CSSProperties['fontWeight']>
     /** Specifies the font-size. */
     fontSize: number
      /** Specifies the line-height. */
@@ -182,22 +182,24 @@ interface SocialButtonTheme {
     paddingX: number
     /** Specifies the padding for the y axis. (top and bottom) */
     paddingY: number
+    /** Specifies the border-color. */
+    borderColor: NonNullable<CSSProperties['color']>,
     /** Specifies the border-radius. */
     borderRadius: number
     /** Specifies the border-width. */
     borderWidth: number
     /** Function that specifies the box shadow based on the border color. */
-    focusBoxShadow: typeof inputBtnFocusBoxShadow
+    focusBoxShadow: (color?: CSSProperties['color']) => NonNullable<CSSProperties['boxShadow']>
     /** Specifies the height. */
     height: number
 }
 
 interface PasswordStrengthTheme {
-    color0: CSS.Color
-    color1: CSS.Color
-    color2: CSS.Color
-    color3: CSS.Color
-    color4: CSS.Color
+    color0: NonNullable<CSSProperties['color']>
+    color1: NonNullable<CSSProperties['color']>
+    color2: NonNullable<CSSProperties['color']>
+    color3: NonNullable<CSSProperties['color']>
+    color4: NonNullable<CSSProperties['color']>
 }
 
 interface Theme extends BaseTheme {
@@ -316,6 +318,12 @@ type LoginViewProps = {
      */
     allowSignup?: boolean;
     /**
+     * Boolean that specifies whether biometric login is enabled.
+     *
+     * @default false
+     */
+    allowWebAuthnLogin?: boolean;
+    /**
      * List of authentication options
      */
     auth?: AuthOptions;
@@ -358,44 +366,6 @@ type LoginViewProps = {
      */
     socialProviders?: string[];
 };
-
-interface LoginWithWebAuthnViewProps {
-    /**
-     * @deprecated
-     */
-    acceptTos?: boolean;
-    /**
-     * Boolean that specifies whether signup is enabled.
-     *
-     * @default true
-     */
-    allowSignup?: boolean;
-    /**
-     * List of authentication options
-     */
-    auth?: AuthOptions;
-    /**
-     * Whether the signup form fields' labels are displayed on the login view.
-     *
-     * @default false
-     */
-    showLabels?: boolean;
-    /**
-     * Lists the available social providers. This is an array of strings.
-     * Tip: If you pass an empty array, social providers will not be displayed.
-     */
-    socialProviders?: string[];
-}
-
-interface LoginWithPasswordViewProps {
-    allowForgotPassword?: boolean;
-    auth?: AuthOptions;
-    canShowPassword?: boolean;
-    recaptcha_enabled?: boolean;
-    recaptcha_site_key?: string;
-    showLabels?: boolean;
-    showRememberMe?: boolean;
-}
 
 interface PasswordSignupFormProps {
     auth?: AuthOptions;
@@ -461,12 +431,6 @@ interface SignupViewProps extends SignupWithPasswordViewProps, SignupWithWebAuth
      */
     allowLogin?: boolean;
     /**
-     * Boolean that specifies whether biometric login is enabled.
-     *
-     * @default false
-     */
-    allowWebAuthnLogin?: boolean;
-    /**
      * Boolean that specifies whether biometric signup is enabled.
      *
      * @default false
@@ -500,12 +464,6 @@ interface ForgotPasswordViewProps {
      */
     showLabels?: boolean;
     /**
-     * Boolean that specifies whether biometric login is enabled.
-     *
-     * @default false
-     */
-    allowWebAuthnLogin?: boolean;
-    /**
      * Boolean that specifies whether reCAPTCHA is enabled or not.
      */
     recaptcha_enabled?: boolean;
@@ -527,16 +485,9 @@ interface ForgotPasswordViewProps {
 }
 interface ForgotPasswordSuccessViewProps {
     allowLogin?: boolean;
-    allowWebAuthnLogin?: boolean;
 }
 
 interface QuickLoginViewProps {
-    /**
-     * Boolean that specifies whether biometric login is enabled.
-     *
-     * @default false
-     */
-    allowWebAuthnLogin?: boolean;
     /**
      * List of authentication options
      */
@@ -601,6 +552,7 @@ interface MainViewProps$5 {
 type FaSelectionViewState = MFA.StepUpResponse;
 type FaSelectionViewProps = Prettify<Partial<MFA.StepUpResponse> & {
     showIntro?: boolean;
+    auth?: AuthOptions;
 }>;
 type StepUpResponse = RequiredProperty<PasswordlessResponse, 'challengeId'>;
 type StepUpHandlerResponse = StepUpResponse & StartPasswordlessFormData;
@@ -614,7 +566,7 @@ type VerificationCodeViewProps$3 = Prettify<Partial<StepUpHandlerResponse> & {
 type MfaStepUpProps = MainViewProps$5 & FaSelectionViewProps & VerificationCodeViewProps$3;
 type MfaStepUpWidgetProps = MfaStepUpProps;
 
-interface AuthWidgetProps extends LoginViewProps, LoginWithWebAuthnViewProps, LoginWithPasswordViewProps, SignupViewProps, SignupWithPasswordViewProps, SignupWithWebAuthnViewProps, ForgotPasswordViewProps, ForgotPasswordSuccessViewProps, QuickLoginViewProps, ReauthViewProps, Omit<FaSelectionViewProps, keyof FaSelectionViewState>, Omit<VerificationCodeViewProps$3, keyof VerificationCodeViewState> {
+interface AuthWidgetProps extends LoginViewProps, SignupViewProps, SignupWithPasswordViewProps, SignupWithWebAuthnViewProps, ForgotPasswordViewProps, ForgotPasswordSuccessViewProps, QuickLoginViewProps, ReauthViewProps, Omit<FaSelectionViewProps, keyof FaSelectionViewState>, Omit<VerificationCodeViewProps$3, keyof VerificationCodeViewState> {
     /**
      * Boolean that specifies whether quick login is enabled.
      *
@@ -626,10 +578,9 @@ interface AuthWidgetProps extends LoginViewProps, LoginWithWebAuthnViewProps, Lo
      *
      * - if `allowLogin` is set to `true`, it defaults to `login`.
      * - if `allowLogin` is set to `false` and `allowSignup` is set to `true`, it defaults to `signup`.
-     * - if `allowLogin` is set to `false` and `allowWebAuthnLogin` is set to `true`, it defaults to `login-with-web-authn`.
      * - otherwise, defaults to `forgot-password`.
      */
-    initialScreen?: 'login' | 'login-with-web-authn' | 'signup' | 'forgot-password';
+    initialScreen?: 'login' | 'signup' | 'forgot-password';
 }
 
 interface MainViewProps$4 {
@@ -1012,7 +963,7 @@ declare class UiClient {
     _ssoCheck<P extends WidgetProps>(widget: Widget<Omit<P, keyof WidgetProps>>, options: P & {
         auth?: AuthOptions;
     }): void;
-    adaptError(error: unknown): string | boolean;
+    adaptError(error: unknown): string;
     handleError(error: unknown): void;
 }